sanicbe/requirements.txt
sanic==21.9.3
asyncio_redis==0.16.0
asyncpg==0.25.0
passlib==1.7.4
pytest==6.2.5
aiohttp==3.8.1
sqlalchemy==1.4.27
contextvars==2.4
sanic-cors==1.0.1
pyjwt==2.3.0
apscheduler==3.8.1
aiofiles==0.8.0
Cerberus==1.3.4
arrow==1.2.1